finish base add
